+++ FlightService/DTOs/AirportReadDto.cs
@@ -2,7 +2,7 @@
 {
     public class AirportReadDto
     {
-        public short AirportId { get; set; }
+        public int AirportId { get; set; }
         public string Iata { get; set; }
         public string Icao { get; set; }
         public string Name { get; set; }
